% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-nn-fit.R
\name{.train_nn_tab_impl}
\alias{.train_nn_tab_impl}
\title{Preprocessing bridge for data.frame and formula methods}
\usage{
.train_nn_tab_impl(
  processed,
  formula,
  hidden_neurons,
  activations,
  output_activation,
  bias,
  arch,
  early_stopping,
  epochs,
  batch_size,
  penalty,
  mixture,
  learn_rate,
  optimizer,
  optimizer_args,
  loss,
  validation_split,
  device,
  verbose,
  cache_weights,
  initial_model = NULL,
  track_optim_path = FALSE
)
}
\description{
Preprocessing bridge for data.frame and formula methods
}
\keyword{internal}
