% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_fun.R
\name{eval_act_funs}
\alias{eval_act_funs}
\title{Activation Function Specs Evaluation}
\usage{
eval_act_funs(activations, output_activation)
}
\arguments{
\item{activations_quo}{Quosure containing the activations expression}

\item{output_activation_quo}{Quosure containing the output_activation expression}
}
\value{
A list with two elements: \code{activations} and \code{output_activation}
}
\description{
Helper function for \code{act_funcs()} argument.
}
\keyword{internal}
