% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_depth.R
\name{extract_depth_param}
\alias{extract_depth_param}
\title{Extract depth parameter values from n_hlayer argument}
\usage{
extract_depth_param(n_hlayer, param_list = list(), levels = 3L)
}
\arguments{
\item{n_hlayer}{Either an integer vector or a param object}

\item{param_list}{List of parameters (for extracting n_hlayers if present)}

\item{levels}{Number of levels for regular grids}
}
\value{
List with \code{values} component containing integer vector of depths
}
\description{
Extract depth parameter values from n_hlayer argument
}
\keyword{internal}
