% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch-code-gen.R
\name{formula_to_expr_transformer}
\alias{formula_to_expr_transformer}
\title{Convert Formula to Expression Transformer}
\usage{
formula_to_expr_transformer(formula_or_fn)
}
\arguments{
\item{formula_or_fn}{A formula like \code{~ .[[1]]} or a function that transforms expressions}
}
\value{
A function that takes an expression and returns a transformed expression, or NULL
}
\description{
Convert Formula to Expression Transformer
}
\keyword{internal}
