% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch-code-gen.R
\name{layer-attributes}
\alias{layer-attributes}
\alias{$.layer_pr}
\title{"Layer" attributes}
\usage{
\method{$}{layer_pr}(x, name)
}
\arguments{
\item{x}{The .layer itself}

\item{name}{It could be the following:
\itemize{
\item \code{i}: Layer index (1-based integer)
\item \code{ind}: Input dimension for the layer
\item \code{out}: Output dimension for the layer
\item \code{is_output}: Logical indicating if this is the output layer
}}
}
\value{
A pronoun, it returns nothing
}
\description{
"Layer" attributes
}
