% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-parsnip.R
\name{make_kindling}
\alias{make_kindling}
\title{Register kindling engines with parsnip}
\usage{
make_kindling()
}
\description{
This function registers the kindling engine for MLP, RNN, and train_nn models
with parsnip. It should be called when the package is loaded.
}
\keyword{internal}
