% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_nn_exec.R
\name{predict.rnn_fit}
\alias{predict.rnn_fit}
\title{Predict Method for RNN Fits}
\usage{
\method{predict}{rnn_fit}(object, newdata, type = "response", ...)
}
\arguments{
\item{object}{An object of class "rnn_fit".}

\item{newdata}{Data frame. New data for predictions.}

\item{type}{Character. Type of prediction: "response" (default) or "prob" for classification.}

\item{...}{Additional arguments (unused).}
}
\description{
Predict Method for RNN Fits
}
