% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-helpers.R
\name{prepare_kindling_args}
\alias{prepare_kindling_args}
\title{Prepare arguments for kindling models}
\usage{
prepare_kindling_args(args)
}
\description{
Prepare arguments for kindling models
}
\keyword{internal}
