% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-nn-exec.R
\name{print.ffnn_fit}
\alias{print.ffnn_fit}
\title{Print method for ffnn_fit objects}
\usage{
\method{print}{ffnn_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "ffnn_fit"}

\item{...}{Additional arguments (unused)}
}
\value{
No return value, called for side effects (printing model summary)
}
\description{
Print method for ffnn_fit objects
}
\keyword{internal}
