% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_arch.R
\name{print.nn_arch}
\alias{print.nn_arch}
\title{Display \code{nn_arch()} configuration}
\usage{
\method{print}{nn_arch}(x, ...)
}
\arguments{
\item{x}{An object of class "nn_arch"}

\item{...}{Additional arguments (unused)}
}
\value{
No return value, just the information
}
\description{
Display \code{nn_arch()} configuration
}
\keyword{internal}
