% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-gen-nn.R
\name{print.nn_fit}
\alias{print.nn_fit}
\title{Print method for nn_fit objects}
\usage{
\method{print}{nn_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"nn_fit"}}

\item{...}{Additional arguments (unused)}
}
\value{
No return value, called for side effects (printing model summary)
}
\description{
Print method for nn_fit objects
}
\keyword{internal}
