% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nn_exec.R
\name{print.rnn_fit}
\alias{print.rnn_fit}
\title{Print method for rnn_fit objects}
\usage{
\method{print}{rnn_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "rnn_fit"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for rnn_fit objects
}
\keyword{internal}
