% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-imp.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{garson}
\alias{olden}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{NeuralNetTools}{\code{\link[NeuralNetTools:garson]{garson()}}, \code{\link[NeuralNetTools:olden]{olden()}}}
}}

