% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.R
\name{rnn_impl}
\alias{rnn_impl}
\title{RNN Implementation}
\usage{
rnn_impl(
  x,
  y,
  hidden_neurons,
  rnn_type = "lstm",
  activations = NULL,
  output_activation = NULL,
  bias = TRUE,
  bidirectional = TRUE,
  dropout = 0,
  epochs = 100,
  batch_size = 32,
  penalty = 0,
  mixture = 0,
  learn_rate = 0.001,
  optimizer = "adam",
  optimizer_args = list(),
  loss = "mse",
  validation_split = 0,
  device = NULL,
  verbose = FALSE,
  cache_weights = FALSE
)
}
\description{
RNN Implementation
}
\keyword{internal}
