% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-nn-fit.R
\name{train_nn_impl}
\alias{train_nn_impl}
\title{Shared core implementation}
\usage{
train_nn_impl(
  x,
  y,
  hidden_neurons,
  activations = NULL,
  output_activation = NULL,
  bias = TRUE,
  arch = NULL,
  early_stopping = NULL,
  epochs = 100,
  batch_size = 32,
  penalty = 0,
  mixture = 0,
  learn_rate = 0.001,
  optimizer = "adam",
  optimizer_args = list(),
  loss = "mse",
  validation_split = 0,
  device = NULL,
  verbose = FALSE,
  cache_weights = FALSE,
  initial_model = NULL,
  track_optim_path = FALSE,
  fit_class = "nn_fit"
)
}
\description{
Shared core implementation
}
\keyword{internal}
