% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized-nn-fitds.R
\name{train_nn_impl_dataset}
\alias{train_nn_impl_dataset}
\title{train_nn implementation for torch datasets}
\usage{
train_nn_impl_dataset(
  dataset,
  no_x,
  no_y,
  is_classification,
  hidden_neurons,
  activations = NULL,
  output_activation = NULL,
  bias = TRUE,
  epochs = 100,
  batch_size = 32,
  penalty = 0,
  mixture = 0,
  learn_rate = 0.001,
  optimizer = "adam",
  optimizer_args = list(),
  loss = "mse",
  validation_split = 0,
  device = NULL,
  verbose = FALSE,
  cache_weights = FALSE,
  arch = NULL,
  fit_class = "nn_fit_ds"
)
}
\description{
train_nn implementation for torch datasets
}
\keyword{internal}
