% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-device.R
\name{validate_device}
\alias{validate_device}
\title{Validate device and get default device}
\usage{
validate_device(device)
}
\arguments{
\item{device}{Character. Requested device.}
}
\value{
Character string of validated device.
}
\description{
Check if requested device is available. And auto-detect available GPU device
or fallback to CPU.
}
\keyword{internal}
