% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-parsnip.R
\name{ffnn_wrapper}
\alias{ffnn_wrapper}
\alias{rnn_wrapper}
\title{Basemodels-tidymodels wrappers}
\usage{
ffnn_wrapper(formula, data, ...)

rnn_wrapper(formula, data, ...)
}
\arguments{
\item{formula}{A formula specifying the model (e.g., \code{y ~ x1 + x2})}

\item{data}{A data frame containing the training data}

\item{...}{Additional arguments passed to the underlying training function}
}
\description{
Basemodels-tidymodels wrappers
}
\section{FFNN (MLP) Wrapper for \code{{tidymodels}} interface}{

This is a function to interface into \code{{tidymodels}}
(do not use this, use \code{kindling::ffnn()} instead).
}

\section{RNN Wrapper for \code{{tidymodels}} interface}{

This is a function to interface into \code{{tidymodels}}
(do not use this, use \code{kindling::rnn()} instead).
}

