% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch-code-gen.R
\name{print.layer_pr}
\alias{print.layer_pr}
\alias{print.layer_index_pr}
\alias{print.layer_input_pr}
\alias{print.layer_output_pr}
\alias{print.layer_is_output_pr}
\title{Print method for the pronouns}
\usage{
\method{print}{layer_pr}(x, ...)

\method{print}{layer_index_pr}(x, ...)

\method{print}{layer_input_pr}(x, ...)

\method{print}{layer_output_pr}(x, ...)

\method{print}{layer_is_output_pr}(x, ...)
}
\arguments{
\item{x}{An object of class "ffnn_fit"}

\item{...}{Additional arguments (unused)}
}
\value{
No return value, prints out the type of pronoun to be used
}
\description{
Print method for the pronouns
}
\section{For \code{.layer}}{

It displays what fields to be accessed by \code{$}.
}

