% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nn_exec.R
\name{print.ffnn_fit}
\alias{print.ffnn_fit}
\title{Print method for ffnn_fit objects}
\usage{
\method{print}{ffnn_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "ffnn_fit"}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for ffnn_fit objects
}
