% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_code_gen.R
\name{substitute_dot}
\alias{substitute_dot}
\title{Recursively Substitute . with Expression}
\usage{
substitute_dot(expr, replacement)
}
\arguments{
\item{expr}{Expression containing \code{.} placeholders}

\item{replacement}{Expression to substitute for \code{.}}
}
\value{
Modified expression
}
\description{
Recursively Substitute . with Expression
}
\keyword{internal}
